requirements.txt
fastapi
uvicorn
SQLAlchemy
passlib[bcrypt]
python-multipart
python-jose
pydantic
pytest
httpx
